Add all errors to one file
