--- lib/inquisition/errors/additional_software_absence_error.rb
@@ -1,9 +1,0 @@
-module Inquisition
-  module Errors
-    class AdditionalSoftwareAbsenceError < BaseError
-      def initialize(command)
-        super("Additional software with command #{command} not found")
-      end
-    end
-  end
-end
