--- lib/inquisition/errors/base_config_absense_error.rb
@@ -1,9 +1,0 @@
-module Inquisition
-  module Errors
-    class BaseConfigAbsenseError < BaseError
-      def initialize
-        super('Run inquisition build first')
-      end
-    end
-  end
-end
