--- lib/inquisition/errors/base_error.rb
@@ -1,6 +1,0 @@
-module Inquisition
-  module Errors
-    class BaseError < StandardError
-    end
-  end
-end
