--- lib/inquisition/errors/invalid_ruby_version_error.rb
@@ -1,9 +1,0 @@
-module Inquisition
-  module Errors
-    class InvalidRubyVersionError < BaseError
-      def initialize(version)
-        super("Invalid ruby version #{version}")
-      end
-    end
-  end
-end
