--- lib/inquisition/errors/no_config_file_error.rb
@@ -1,9 +1,0 @@
-module Inquisition
-  module Errors
-    class NoConfigFileError < BaseError
-      def initialize(name)
-        super("File #{name} not found")
-      end
-    end
-  end
-end
