--- lib/inquisition/errors/no_gem_error.rb
@@ -1,9 +1,0 @@
-module Inquisition
-  module Errors
-    class NoGemError < BaseError
-      def initialize(name)
-        super("Gem #{name} not found")
-      end
-    end
-  end
-end
